resume mode and initial index options

+++ scripts/openduck.py
@@ -161,6 +161,8 @@ def args_sanitation(parser, modes):
                 if 'threads' in input_arguments: args.threads = int(input_arguments['threads'])
                 if 'keep_all_files' in input_arguments: args.keep_all_files = bool(input_arguments['keep_all_files'])
                 if 'fix_ligand' in input_arguments : args.fix_ligand = bool(input_arguments['fix_ligand'])
+                if 'resume' in input_arguments : args.resume = bool(input_arguments['resume'])
+                if 'i0' in input_arguments : args.i0 = int(input_arguments['i0'])
                 if args.queue_template == 'local' and args.batch: args.queue_template = None # no local array script
                 if (not args.ligand.endswith('.sdf') and not args.ligand.endswith('.sd')) and args.batch:
                     modes.choices['amber-prepare'].error('Batch processing requires the ligand to be in SD or SDF format.')
@@ -322,6 +324,8 @@ def parse_input():
     amber_prep.add_argument('-t', '--threads', default=1, type=int, help='Define number of CPUs for batch processing.')
     amber_prep.add_argument('-fl','--fix-ligand', action='store_true', help='Some simple fixes for the ligand: ensure tetravalent nitrogens have the right charge assigned and add missing hydrogen atoms.')
     amber_prep.add_argument('--keep-all-files', default=False, action='store_true', help='Disable cleaning up intermediate files during preparation and simulations.')
+    amber_prep.add_argument('--resume', default=False, action='store_true', help='Enable the resume mode. Protecting LIG_target folders already prepared and starting from the not done, avoiding overwritting.')
+    amber_prep.add_argument('i0', '--index0', default=1, type=int, help='Starting index for naming batch ligands. Default: 1.')
     #Arguments for report
     report = modes.add_parser('report', help='Generate a report for OpenDUck results.', description='Generate a table report for dynamic undocking output. For a multi-ligand report, use the pattern flag with wildcards to the directories.')
     report.set_defaults(mode='Report')
@@ -501,7 +505,7 @@ def prepare_sys_for_amber(ligand_file, protein_file, chunk_file, interaction, HM
     AMBER = Amber_templates(structure=p[0], interaction=p[1:],hmr=HMR, seed=seed)
     AMBER.write_all_inputs()
 
-def AMBER_prepare_ligand_in_folder(ligand_string, lig_indx, protein, chunk, interaction, HMR, base_dir, small_molecule_forcefield = 'SMIRNOFF', water_model = 'tip3p', forcefield = 'amber99sb', ion_strength = 0.1, box_buffer_distance = 10, waters_to_retain='waters_to_retain.pdb', seed='-1', fix_ligand=False, clean_up=False):
+def AMBER_prepare_ligand_in_folder(ligand_string, lig_indx, protein, chunk, interaction, HMR, base_dir, small_molecule_forcefield = 'SMIRNOFF', water_model = 'tip3p', forcefield = 'amber99sb', ion_strength = 0.1, box_buffer_distance = 10, waters_to_retain='waters_to_retain.pdb', seed='-1', fix_ligand=False, clean_up=False, resume=False):
     '''
     Generate the folder for a ligand preparation and prepare such ligand.
     '''
@@ -511,9 +515,12 @@ def AMBER_prepare_ligand_in_folder(ligand_string, lig_indx, protein, chunk, inte
     os.chdir(base_dir)
 
     #Create the ligand folder
-    if os.path.isdir(f'LIG_target_{lig_indx}'):
+    if os.path.isdir(f'LIG_target_{lig_indx}') and not resume:
         print(f'WARNING: LIG_target_{lig_indx} already exist and it will be overwritten.')
         shutil.rmtree(f'./LIG_target_{lig_indx}', ignore_errors=True)
+    elif os.path.isdir(f'LIG_target_{lig_indx}') and resume:
+        print(f'WARNING: LIG_target_{lig_indx} already exist and will be skipped.')
+        return(f'Lig_target_{lig_indx} skipped.')
     os.mkdir(f'LIG_target_{lig_indx}')
     os.chdir(f'LIG_target_{lig_indx}')
     print(f'Working on LIG_target_{lig_indx}')
@@ -641,10 +648,10 @@ def do_AMBER_preparation(args):
         base_dir = os.getcwd()
 
         r = [pool.apply_async(AMBER_prepare_ligand_in_folder,
-                          args=(ligand_string, j+1, args.receptor, chunk_file,
+                          args=(ligand_string, j+args.i0, args.receptor, chunk_file,
                                 args.interaction, args.HMR, base_dir,
                                 args.small_molecule_forcefield, args.water_model, args.protein_forcefield,
-                                args.ionic_strength, args.solvent_buffer_distance, args.waters_to_retain, args.seed, args.fix_ligand, not args.keep_all_files),
+                                args.ionic_strength, args.solvent_buffer_distance, args.waters_to_retain, args.seed, args.fix_ligand, not args.keep_all_files, args.resume),
                           callback=log_result,
                           error_callback=handle_error) for j, ligand_string in enumerate(ligand_string_generator(args.ligand))]
         pool.close()
